+++ librustic/src/communication/uci.rs
@@ -2,8 +2,8 @@ pub mod cmd_in;
 pub mod cmd_out;
 pub mod defs;
 pub mod init;
-pub mod output;
 pub mod parse;
+pub mod print;
 pub mod thread_in;
 pub mod thread_out;
